pyproject.toml
[tool.poetry]
name = "grocery-list-app"
version = "0.1.0"
description = ""
authors = ["Helena <heleribera@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.9.7 || >3.9.7,<4.0"
streamlit = "^1.16.0"
pandas = "^2.0.0"
seaborn = "^0.12.2"
sqlalchemy = "^2.0.9"
pydantic = "^1.10.7"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"
streamlit-autorefresh = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
